Moved callbacks to it's own directory. Removed sub-classing of standard pl callbacks. Added extra parameters to config for pl's callbacks.
