+++ README.md
@@ -97,6 +97,29 @@ It is also used to save the config files for each run.
 
 Override or modify the `_log_images_to_wandb()` and `_save_results_to_disk()` methods for your application.
 
+#### EarlyStopping
+
+Here's some tips on how to configure early stopping:
+
+```
+Args:
+        monitor: Monitor a key validation metric (eg: mIoU). Monitoring loss is not a good idea as it is an unreliable
+                 indicator of model performance. Two models might have the same loss but different performance
+                 or the loss might start increasing, even though performance does not decrease.
+
+        min_delta: Project-dependent - choose a value for your metric below which you'd consider the improvement
+                   negligible.
+                   Example: For segmentation, I do not care for improvements less than 0.05% IoU in general.
+                            But in kaggle competitions, even 0.01% would matter.
+
+        patience: Patience is the number of val epochs to wait for to see an improvement. It is affected by the
+                  ``check_val_every_n_epoch`` and ``val_check_interval`` params to the PL Trainer.
+
+                  Takes experimentation to figure out appropriate patience for your project. Train the model
+                  once without early stopping and see how long it takes to converge on a given dataset.
+                  Choose the number of epochs between when you feel it's started to converge and after you're
+                  sure the model has converged. Reduce the patience if you see the model continues to train for too long.
+```
 
 ### Notes
 #### Absolute imports
