+++ seg_lapa/config_parse/callbacks_available.py
@@ -2,16 +2,18 @@
 from typing import Optional
 
 from pydantic.dataclasses import dataclass
+from pytorch_lightning.callbacks import Callback, ModelCheckpoint, EarlyStopping
 
 from seg_lapa.config_parse.conf_utils import asdict_filtered
-from pytorch_lightning.callbacks import Callback, ModelCheckpoint
-from seg_lapa.callbacks import EarlyStopping, LogMedia
+from seg_lapa.callbacks.log_media import LogMedia
 
 
 @dataclass(frozen=True)
 class EarlyStopConf:
+    monitor: str
     min_delta: float
     patience: int
+    mode: str
 
     def get_callback(self) -> Callback:
         return EarlyStopping(**asdict_filtered(self))
@@ -23,6 +25,7 @@ class CheckpointConf:
     mode: str
     save_last: Optional[bool]
     period: int
+    save_top_k: Optional[int]
 
     def get_callback(self, logs_dir) -> Callback:
         return ModelCheckpoint(dirpath=logs_dir, **asdict_filtered(self))
